Add signup in AuthenticationService
Move Credentials and LoginContext to interfaces
