+++ src/app/core/authentication/authentication.service.ts
@@ -1,21 +1,15 @@
 import { Injectable } from '@angular/core';
-import { HttpClient, HttpHeaders, HttpErrorResponse } from '@angular/common/http';
-import { Observable, of, throwError } from 'rxjs';
-import { catchError, map } from 'rxjs/operators';
-
-export interface Credentials {
-  username: string;
-  token: string;
-}
-
-export interface LoginContext {
-  username: string;
-  password: string;
-  remember?: boolean;
-}
+import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { Observable, of } from 'rxjs';
+import { map } from 'rxjs/operators';
+import { Credentials, LoginContext } from '@app/shared/interfaces';
 
 const credentialsKey = 'credentials';
-const authUrl = 'http://localhost:3780/auth';
+
+const routes = {
+  login: () => `auth/login`,
+  signup: () => `auth/signup`,
+};
 
 /**
  * Provides a service for the authentication workflow.
@@ -46,11 +40,32 @@ export class AuthenticationService {
       })
     };
 
-    return this.http.get<Credentials>(authUrl + '/login', httpOptions).pipe(
+    return this.http.get<Credentials>(routes.login(), httpOptions).pipe(
       map(response => {
         credentials = response;
-        this.setCredentials(credentials, context.remember);
+        this.setCredentials(credentials, true);
         return credentials;
+      })
+    );
+  }
+
+  /**
+   * Signs up and authenticates the user.
+   * @param context The login parameters.
+   * @return The username.
+   */
+  signup(context: LoginContext): Observable<string> {
+    const httpOptions = {
+      headers: new HttpHeaders({
+        'Content-Type': 'application/json'
+      })
+    };
+
+    const signupContext = { username: context.username, password: context.password };
+
+    return this.http.post<any>(routes.signup(), signupContext, httpOptions).pipe(
+      map(response => {
+        return response.username;
       })
     );
   }
